refactor(assets): Move things around to avoid importing image service unnecessarily
