+++ packages/astro/src/assets/endpoint/node.ts
@@ -3,10 +3,11 @@ import { readFile } from 'fs/promises';
 import mime from 'mime/lite.js';
 import os from 'os';
 import type { APIRoute } from '../../@types/astro.js';
-import { getConfiguredImageService, isRemoteAllowed } from '../internal.js';
+import { getConfiguredImageService } from '../internal.js';
 import { etag } from '../utils/etag.js';
 // @ts-expect-error
 import { assetsDir, imageConfig } from 'astro:assets';
+import { isRemoteAllowed } from '../utils/remotePattern.js';
 
 function replaceFileSystemReferences(src: string) {
 	return os.platform().includes('win32') ? src.replace(/^\/@fs\//, '') : src.replace(/^\/@fs/, '');
