+++ packages/astro/src/assets/utils/index.ts
@@ -1,4 +1,14 @@
 export { emitESMImage } from './emitAsset.js';
+export { isESMImportedImage, isRemoteImage } from './imageKind.js';
 export { imageMetadata } from './metadata.js';
 export { getOrigQueryParams } from './queryParams.js';
+export {
+	isRemoteAllowed,
+	matchHostname,
+	matchPathname,
+	matchPattern,
+	matchPort,
+	matchProtocol,
+	type RemotePattern,
+} from './remotePattern.js';
 export { hashTransform, propsToFilename } from './transformToPath.js';
