+++ packages/astro/src/assets/utils/transformToPath.ts
@@ -2,8 +2,8 @@ import { deterministicString } from 'deterministic-object-hash';
 import { basename, extname } from 'node:path';
 import { removeQueryString } from '../../core/path.js';
 import { shorthash } from '../../runtime/server/shorthash.js';
-import { isESMImportedImage } from '../internal.js';
 import type { ImageTransform } from '../types.js';
+import { isESMImportedImage } from './imageKind.js';
 
 export function propsToFilename(transform: ImageTransform, hash: string) {
 	let filename = removeQueryString(
